requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
pyarrow>=14.0.0
numpy>=1.26.0
scikit-learn>=1.4.0
umap-learn>=0.5.6
hdbscan>=0.8.33
plotly>=5.18.0
pydantic>=2.5.0
joblib>=1.3.0
diskcache>=5.6.3
sentence-transformers>=2.3.1
openai>=1.12.0
pyyaml>=6.0.1
httpx>=0.26.0
tenacity>=8.2.3
python-dotenv>=1.0.0
